SSH agent improvements, fix python LSP behavior in nvim
Setup the 1pass SSH agent so that it's only enabled on machines with 1pass and disabled when the machine is accessed via SSH.

Fix the navigator.lua plugin so that python LSPs and formatters load correctly. Yapf is disabled for now though since it doesn't
respect the style file in repos yet.